fix(bugs): 修复了一些已知的BUG

+++ OllamaK/view/MessageListView.swift
@@ -5,8 +5,8 @@
 //  Created by Hamm on 2024/11/13.
 //
 
-import SwiftUI
 import Parma
+import SwiftUI
 
 struct MessageListView: View {
     /*
@@ -15,9 +15,10 @@ struct MessageListView: View {
     @Binding public var messages: [OllamaMessage]
 
     var body: some View {
-        ForEach($messages.filter({ item in
-            return OllamaMessage.ROLE_SYSTEM != item.role.wrappedValue
-        }), id: \.id
+        ForEach(
+            $messages.filter({ item in
+                return OllamaMessage.ROLE_SYSTEM != item.role.wrappedValue
+            }), id: \.id
         ) { $message in
             MessageItem(message: message)
         }.padding(0)
@@ -28,41 +29,72 @@ struct MessageListView: View {
      */
     private func MessageItem(message: OllamaMessage) -> some View {
         let config = getMessageItemConfig(role: message.role)
+        let content = message.content.isEmpty ? "." : message.content
         return HStack {
             HStack {
                 if message.role != OllamaMessage.ROLE_ASSISTANT {
                     Spacer()
                 }
-                Parma((message.content.isEmpty ? "." : message.content))
-                    .font(
-                        .system(
-                            size: config.fontSize)
-                    )
-                    .padding(12)
-                    .foregroundColor(.gray)
-                    .background(
-                        RoundedRectangle(cornerRadius: 12)
-                            .fill(
-                                config.bgColor
-                                    .opacity(0.1))
-                    )
-                    .padding(.leading, config.paddingLeft)
-                    .padding(.trailing, config.paddingRight)
-                    .padding(.vertical, config.verticalPadding)
-                    .contextMenu {
-                        Button {
-                            deleteMessage(item: message)
-                        } label: {
-                            Label("删除这条消息", systemImage: "trash")
-                        }
-                    }
                 if message.role != OllamaMessage.ROLE_USER {
+                    Parma(content)
+                        .font(
+                            .system(
+                                size: config.fontSize)
+                        )
+                        .padding(12)
+                        .foregroundColor(.gray)
+                        .background(
+                            RoundedRectangle(cornerRadius: 12)
+                                .fill(
+                                    config.bgColor
+                                        .opacity(0.1))
+                        )
+                        .padding(.leading, config.paddingLeft)
+                        .padding(.trailing, config.paddingRight)
+                        .padding(.vertical, config.verticalPadding)
+                        .contextMenu {
+                            Button {
+                                copyMessage(item: message)
+                            } label: {
+                                Label("复制内容", systemImage: "document.on.document")
+                            }
+                            Button {
+                                deleteMessage(item: message)
+                            } label: {
+                                Label("删除这条消息", systemImage: "trash")
+                            }
+                        }
                     Spacer()
+                }
+                if message.role == OllamaMessage.ROLE_USER {
+                    Text(content)
+                        .font(
+                            .system(
+                                size: config.fontSize)
+                        )
+                        .padding(12)
+                        .foregroundColor(.gray)
+                        .background(
+                            RoundedRectangle(cornerRadius: 12)
+                                .fill(
+                                    config.bgColor
+                                        .opacity(0.1))
+                        )
+                        .padding(.leading, config.paddingLeft)
+                        .padding(.trailing, config.paddingRight)
+                        .padding(.vertical, config.verticalPadding)
+                        .contextMenu {
+                            Button {
+                                deleteMessage(item: message)
+                            } label: {
+                                Label("删除这条消息", systemImage: "trash")
+                            }
+                        }
                 }
             }
         }
     }
-    
+
     /*
      删除指定索引的历史
      */
@@ -74,6 +106,14 @@ struct MessageListView: View {
     }
 
     /*
+     复制消息
+     */
+    private func copyMessage(item: OllamaMessage) {
+        let pasteboard = UIPasteboard.general
+        pasteboard.string = item.content
+    }
+
+        /*
      获取消息配置
      */
     private func getMessageItemConfig(role: String) -> MessageItemConfig {
